+++ src/pwm.rs
@@ -245,7 +245,10 @@ macro_rules! hal {
                 }
 
                 fn get_duty(&self) -> u16 {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr1.read().ccr().bits() }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr1.read().ccr1().bits() }
                 }
 
                 fn get_max_duty(&self) -> u16 {
@@ -253,7 +256,10 @@ macro_rules! hal {
                 }
 
                 fn set_duty(&mut self, duty: u16) {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr1.write(|w| w.ccr().bits(duty)) }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr1.write(|w| w.ccr1().bits(duty)) }
                 }
             }
 
@@ -269,7 +275,10 @@ macro_rules! hal {
                 }
 
                 fn get_duty(&self) -> u16 {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr2.read().ccr().bits() }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr2.read().ccr2().bits() }
                 }
 
                 fn get_max_duty(&self) -> u16 {
@@ -277,7 +286,10 @@ macro_rules! hal {
                 }
 
                 fn set_duty(&mut self, duty: u16) {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr2.write(|w| w.ccr().bits(duty)) }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr2.write(|w| w.ccr2().bits(duty)) }
                 }
             }
 
@@ -293,7 +305,10 @@ macro_rules! hal {
                 }
 
                 fn get_duty(&self) -> u16 {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr3.read().ccr().bits() }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr3.read().ccr3().bits() }
                 }
 
                 fn get_max_duty(&self) -> u16 {
@@ -301,7 +316,10 @@ macro_rules! hal {
                 }
 
                 fn set_duty(&mut self, duty: u16) {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr3.write(|w| w.ccr().bits(duty)) }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr3.write(|w| w.ccr3().bits(duty)) }
                 }
             }
 
@@ -317,7 +335,10 @@ macro_rules! hal {
                 }
 
                 fn get_duty(&self) -> u16 {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr4.read().ccr().bits() }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr4.read().ccr4().bits() }
                 }
 
                 fn get_max_duty(&self) -> u16 {
@@ -325,7 +346,10 @@ macro_rules! hal {
                 }
 
                 fn set_duty(&mut self, duty: u16) {
+                    #[cfg(feature = "stm32f103")]
                     unsafe { (*$TIMX::ptr()).ccr4.write(|w| w.ccr().bits(duty)) }
+                    #[cfg(not(feature = "stm32f103"))]
+                    unsafe { (*$TIMX::ptr()).ccr4.write(|w| w.ccr4().bits(duty)) }
                 }
             }
         )+